backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pytest
pytest-asyncio
httpx
email-validator 